Improve peak detection by better rejecting invalid peaks
We now move a window over all the peaks and take the sum. The window with the highest sum (the highest average peaks) is then most likely the window that matches the data.

This seems to work really well. Most of our errors are now bit flips instead of missing a peak and shifting all bits left or right.